+++ src/gui/mod.rs
@@ -2010,6 +2010,7 @@ pub enum AuraColumnKey {
     LocalVersion,
     RemoteVersion,
     Author,
+    Type,
     Status,
 }
 
@@ -2022,6 +2023,7 @@ impl AuraColumnKey {
             LocalVersion => localized_string("local"),
             RemoteVersion => localized_string("remote"),
             Author => localized_string("author"),
+            Type => localized_string("type"),
             Status => localized_string("status"),
         }
     }
@@ -2034,6 +2036,7 @@ impl AuraColumnKey {
             LocalVersion => "local",
             RemoteVersion => "remote",
             Author => "author",
+            Type => "type",
             Status => "status",
         };
 
@@ -2048,6 +2051,7 @@ impl From<&str> for AuraColumnKey {
             "local" => AuraColumnKey::LocalVersion,
             "remote" => AuraColumnKey::RemoteVersion,
             "author" => AuraColumnKey::Author,
+            "type" => AuraColumnKey::Type,
             "status" => AuraColumnKey::Status,
             _ => panic!(format!("Unknown AuraColumnKey for {}", s)),
         }
@@ -2097,6 +2101,12 @@ impl Default for AuraHeaderState {
                 },
                 AuraColumnState {
                     key: AuraColumnKey::Author,
+                    btn_state: Default::default(),
+                    width: Length::Units(85),
+                    hidden: false,
+                },
+                AuraColumnState {
+                    key: AuraColumnKey::Type,
                     btn_state: Default::default(),
                     width: Length::Units(85),
                     hidden: false,